+++ src/cli/Microsoft.DotNet.UpgradeAssistant.Cli/Commands/Analyze/ConsoleAnalyzeCommand.cs
@@ -1,7 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.CommandLine;
 using System.CommandLine.Invocation;
 using System.CommandLine.Parsing;
 using System.Threading;
@@ -10,17 +9,17 @@ using Microsoft.Extensions.Hosting;
 
 namespace Microsoft.DotNet.UpgradeAssistant.Cli
 {
-    public delegate IHostBuilder CreateConsoleHost(UpgradeOptions options, ParseResult parseResult);
-
     public class ConsoleAnalyzeCommand : UpgradeAssistantCommand
     {
-        public ConsoleAnalyzeCommand(CreateConsoleHost createHost)
+        public ConsoleAnalyzeCommand()
             : base("analyze")
         {
             IsHidden = true;
-            Handler = CommandHandler.Create<ParseResult, UpgradeOptions, CancellationToken>((result, options, token) => createHost(options, result)
-                .ConfigureServices(ConfigureService)
-                .RunUpgradeAssistantAsync(token));
+            Handler = CommandHandler.Create<ParseResult, UpgradeOptions, CancellationToken>((result, options, token) =>
+                Host.CreateDefaultBuilder()
+                    .UseConsoleUpgradeAssistant(options, result)
+                    .ConfigureServices(ConfigureService)
+                    .RunUpgradeAssistantAsync(token));
         }
 
         public static void ConfigureService(IServiceCollection services)
